Added major requirement fetcher, fixed get_courses
Major requirement fetcher allows user to get a tuple of courses
required to graduate. The first item is course title and second item
is a link to the course's catalog page. Interchangeable courses are
stored in a list containing multiple tuples. The last number of the list
is the number of courses from the list a student needs to take. Also
fixed the get_courses_offered function to get ALL courses offered
instead of only the courses on the first page. Currently working on
setting the prerequisites and creating nodes for each course.