Added Series Grouping Fixed Duplication Bug
I modified get_major_reqs to group courses by series and also all of
the interchangeable series into a single list. Currently the function
returns a tuple with course name and a link to the course page -> I
might change this to a dictionary that way function is cleaner.
Currently need to work on parsing the requisites that the get_class_req
function grabs and linking each course to their prerequisites.